Moved widget components to a single parent folder
