+++ app/javascript/components/dashboard-widgets/dashboard-charts/emptyChart.js
