+++ app/javascript/components/dashboard-widgets/widget-chart/index.jsx
