+++ app/javascript/components/dashboard-widgets/widget-error/index.jsx
