+++ app/javascript/components/dashboard-widgets/widget-footer/index.jsx
