+++ app/javascript/components/dashboard-widgets/widget-report/index.jsx
