+++ app/javascript/components/dashboard-widgets/widget-zoom/index.jsx
