+++ app/javascript/packs/component-definitions-common.js
@@ -23,7 +23,7 @@ import ProviderForm from '../components/provider-form';
 import ChargebackRate from '../components/data-tables/chargeback-rate';
 import CopyCatalogForm from '../components/copy-catalog-form/copy-catalog-form';
 import CopyDashboardForm from '../components/copy-dashboard-form/copy-dashboard-form';
-import DashboardWidget from '../components/dashboard-charts';
+import DashboardWidget from '../components/dashboard-widgets/dashboard-charts';
 import Datastore from '../components/data-tables/datastore';
 import DbList from '../components/data-tables/db-list';
 import FlavorForm from '../components/flavor-form';
@@ -103,13 +103,13 @@ import UtilizationChartGraph from '../components/provider-dashboard-charts/provi
 import VmCommonRenameForm from '../components/vm-common-rename-form';
 import VmServerRelationshipForm from '../components/vm-server-relationship-form';
 import VmSnapshotForm from '../components/vm-snapshot-form/vm-snapshot-form';
-import WidgetChart from '../components/widget-chart';
-import WidgetError from '../components/widget-error';
-import WidgetFooter from '../components/widget-footer';
-import WidgetMenu from '../components/widget-menu';
-import WidgetReport from '../components/widget-report';
-import WidgetWrapper from '../components/widget-wrapper';
-import WidgetZoom from '../components/widget-zoom';
+import WidgetChart from '../components/dashboard-widgets/widget-chart';
+import WidgetError from '../components/dashboard-widgets/widget-error';
+import WidgetFooter from '../components/dashboard-widgets/widget-footer';
+import WidgetMenu from '../components/dashboard-widgets/widget-menu';
+import WidgetReport from '../components/dashboard-widgets/widget-report';
+import WidgetWrapper from '../components/dashboard-widgets/widget-wrapper';
+import WidgetZoom from '../components/dashboard-widgets/widget-zoom';
 import WorkersForm from '../components/workers-form/workers-form';
 import PhysicalStorageForm from '../components/physical-storage-form';
 import SearchBar from '../components/search-bar';
